submitted
tried to do the additional challenge. 
ran into issues doing it on my own and referenced a java implementation on stack overflow at some point, but that implementation had it's own issues and used an array rather than a linked list. Hope this is in the ballpark of what is needed... all of the tests past on my IDE with just getting rid of the "que full" exception raise.